Fix export sundivided polygon meshes
Previous method use as low-level polygonal data, as subdivided data. That's why the export process is crashed. Now it always use only subdivided data. As a consequence, renderer polygonmesh can contains wrong triangulation of concave polygons. Because there is no simple method in the Softimage to get proper triangulation of the polygons. Corresponding data contained in low-level data (PolygonMesh class) and can not be used in our case.