Implement baking
But bake margin image pipline plugon does not work. The result is the same with or without it. Also denoising in the bake image is not working.

We does not save baking result, because the renderer do it himself by using baking properties (which is different from film output options)